Fix saving issue on PPK1.
Also squash functions that do more or less the same, rename files plus
some other small refactorings.
